--- server/requirements.txt
@@ -1,1 +1,0 @@
-supabase == 2.18.1
